requirements.txt
pytesseract
pillow
gTTS
playsound==1.2.2
pyttsx3
mutagen
moviepy
